// cypress/types/IUIUtil.ts
export interface IUIUtil {
    checkCheckbox: (label: string) => void;
    checkRadio: (label: string) => void;
    clearInputField: (label: string) => void; // todo
    clickBreadCrumLink: (text: string) => void // todo
    clickButton: (label: string) => void;
    clickDropdownAndSelectOption: (text: string, option: string) => void; // todo
    clickInputBox: (label: string) => void;
    clickInputButton: (label: string) => void;
    clickLabel: (label: string) => void;
    clickLink: (label: string) => void;
    clickListItem: (value: string) => void;
    clickRadioOption: (label: string) => void; // todo
    inputDate: (label: string, text: string) => void;
    inputText: (label: string, text: string, family?: number) => void;
    inputTextArea: (label: string, text: string) => void;
    inputTextDynamic: (label: string, obj: string, text: string) => void;
    selectByValue: (label: string, option: string, family?: number) => void;
    unCheckCheckbox: (label: string) => void;
    unCheckRadio: (label: string) => void;
    verifyCaptionStyleTextExists: (text: string) => void;
    verifyCaptionStyleTextNotExists: (text: string) => void;
    verifyDropdownExists: (label: string) => void; //todo
    verifyDropdownValue: (label: string, text: string) => void; //todo
    verifyDropdownValueTypeButton: (label: string, text: string) => void; //todo
    verifyErrorNotification: (text: string) => void;
    verifyFieldInLineError: (label: string, error: any) => void;
    verifyFieldInLineMessage: (label: string, error: any) => void; //todo
    verifyFilterComponents: (text: string) => void; //todo
    verifyHeaderExists: (text: string) => void;
    verifyInfoNotification: (text: string) => void; //todo
    verifyInputBoxExists: (label: string) => void; //todo
    verifyInputBoxNotExists: (label: string) => void; //todo
    verifyInputBoxText: (label: string, text: string) => void; //todo
    verifyLabelText: (labelName: string) => void; //todo
    verifyLeavePagePopupNotPresent: () => void; //todo
    verifyLinkExists: (text: string) => void; //todo
    verifyListItem: (value: string) => void;
    verifyNotificationNotExists: (text: string) => void;
    verifyNotificationTextExists: (text: string) => void;
    verifyPageErrors: (...error: any) => void; // todo
    verifyParagraphTextExists: (text: string) => void;
    verifyParagraphTextNotExists: (text: string) => void;
    verifySubHeaderExists: (text: string) => void
    verifySuccessNotification: () => void; //todo
    verifyTableHeader: (headerList: string) => void; //todo
    verifyThirdHeading: (text: string) => void; //todo
}